Crate saluki_env

Helpers for querying environment-specific data.

In many cases, a Saluki service will need to interact with the environment in which it is running. This may include interacting with the operating system, introspecting workloads running on the host, and so on.

This crate provides implementations of various “providers” – specific facets of the environment that can be queried – for a number of different environments, such as bare metal hosts, Kubernetes clusters, and so on.

Re-exports

pub use self::autodiscovery::AutodiscoveryProvider;

pub use self::host::HostProvider;

pub use self::workload::WorkloadProvider;

Modules

autodiscovery

Autodiscovery provider.

configstream

Config stream.

features

Feature detection.

helpers

Miscellaneous helpers used by different providers.

host

Host provider.

workload

Workload provider.

Traits

EnvironmentProvider

Provides information about the environment in which the process is running.