saluki_env::features

Struct Feature

pub struct Feature { /* private fields */ }

Features are distinct markers or indicators of a particular technology or platform that is present.

In general, features represent the type of environment that the application is running in, such as the Kubernetes feature being present indicating that the application is running in Kubernetes, and so on.

Implementations

impl Feature

pub const HostMappedProcfs: Feature

Host-mapped procfs.

This implies that we’re in a containerized environment and the host’s procfs (/proc) has been mapped into the container using a /host prefix, resulting in a /host/proc path.

pub const HostMappedCgroupfs: Feature

Host-mapped cgroupfs.

This implies that we’re in a containerized environment and the host’s cgroupfs (/sys/fs/cgroup) has been mapped into the container using a /host prefix, resulting in a /host/sys/fs/cgroup path.

pub const Containerd: Feature

Containerd.

pub const fn bits(&self) -> u16

Returns the underlying bits of the bitmask.

pub const fn all_bits() -> Self

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn all_flags() -> Self

Returns a bitmask that contains all flags.

pub const fn is_all_bits(&self) -> bool

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn is_all_flags(&self) -> bool

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn all() -> Self

👎Deprecated: Please use the ::all_bits() constructor

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn is_all(&self) -> bool

👎Deprecated: Please use the .is_all_bits() method

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn full() -> Self

👎Deprecated: Please use the ::all_flags() constructor

Returns a bitmask that contains all flags.

pub const fn is_full(&self) -> bool

👎Deprecated: Please use the .is_all_flags() method

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn none() -> Self

Returns a bitmask that does not contain any values.

pub const fn is_none(&self) -> bool

Returns true if the bitmask does not contain any values.

pub const fn truncate(&self) -> Self

Returns a bitmask that only has bits corresponding to flags

pub const fn intersects(&self, other: Self) -> bool

Returns true if self intersects with any value in other, or if other does not contain any values.

This is equivalent to (self & other) != 0 || other == 0.

pub const fn contains(&self, other: Self) -> bool

Returns true if self contains all values of other.

This is equivalent to (self & other) == other.

pub const fn not(self) -> Self

Returns the bitwise NOT of the bitmask.

pub const fn and(self, other: Self) -> Self

Returns the bitwise AND of the bitmask.

pub const fn or(self, other: Self) -> Self

Returns the bitwise OR of the bitmask.

pub const fn xor(self, other: Self) -> Self

Returns the bitwise XOR of the bitmask.