saluki_env::helpers::remote_agent

Struct RemoteAgentClient

pub struct RemoteAgentClient { /* private fields */ }

A client for interacting with the Datadog Agentโs internal gRPC-based API.

Implementations

impl RemoteAgentClient

pub async fn from_configuration( config: &GenericConfiguration, ) -> Result<Self, GenericError>

Creates a new RemoteAgentClient from the given configuration.

Errors

If the Agent gRPC client cannot be created (invalid API endpoint, missing authentication token, etc), or if the authentication token is invalid, an error will be returned.

pub async fn get_hostname(&mut self) -> Result<String, GenericError>

Gets the detected hostname from the Agent.

Errors

If there is an error querying the Agent API, an error will be returned.

pub fn get_tagger_stream( &mut self, cardinality: TagCardinality, ) -> StreamingResponse<StreamTagsResponse>

Gets a stream of tagger entities at the given cardinality.

If there is an error with the initial request, or an error occurs while streaming, the next message in the stream will be Some(Err(status)), where the status indicates the underlying error.

pub fn get_workloadmeta_stream( &mut self, ) -> StreamingResponse<WorkloadmetaStreamResponse>

Gets a stream of all workloadmeta entities.

If there is an error with the initial request, or an error occurs while streaming, the next message in the stream will be Some(Err(status)), where the status indicates the underlying error.

pub async fn register_remote_agent_request( &mut self, pid: u32, display_name: &str, api_endpoint: &str, services: Vec<String>, ) -> Result<Response<RegisterRemoteAgentResponse>, GenericError>

Registers a Remote Agent with the Agent.

Errors

If there is an error sending the request to the Agent API, an error will be returned.

pub async fn refresh_remote_agent_request( &mut self, session_id: &str, ) -> Result<Response<RefreshRemoteAgentResponse>, GenericError>

Refreshes a Remote Agent with the Agent.

If there is an error sending the request to the Agent API, an error will be returned.

pub async fn get_host_tags( &self, ) -> Result<Response<HostTagReply>, GenericError>

Gets the host tags from the Agent.

Errors

If there is an error querying the Agent API, an error will be returned.

pub fn get_autodiscovery_stream( &mut self, ) -> StreamingResponse<AutodiscoveryStreamResponse>

Gets a stream of autodiscovery config updates.

If there is an error with the initial request, or an error occurs while streaming, the next message in the stream will be Some(Err(status)), where the status indicates the underlying error.

pub fn stream_config_events(&mut self) -> StreamingResponse<ConfigEvent>

Gets a stream of config events.

If there is an error with the initial request, or an error occurs while streaming, the next message in the stream will be Some(Err(status)), where the status indicates the underlying error.