saluki_env::workload::entity

Struct HighestPrecedenceEntityIdRef

pub struct HighestPrecedenceEntityIdRef<'a>(/* private fields */);

A wrapper for entity IDs that sorts them in a manner consistent with the expected precedence of entity IDs.

This type establishes a total ordering over entity IDs based on their logical precedence, which is as follows:

global (highest precedence)

container

OCI image metadata

ECS task

Kubernetes deployment, metadata, node, and pod

process

local container inode and PID aliases (lowest precedence)

Wrapped entity IDs are be sorted highest to lowest precedence. For entity IDs with the same precedence, they’re further ordered by their internal value. For entity IDs with a string identifier, lexicographical ordering is used. For entity IDs with a numeric identifier, numerical ordering is used.