saluki_env::workload

Enum EntityId

pub enum EntityId {
    Global,
    PodUid(MetaString),
    Container(MetaString),
    ContainerInode(u64),
    ContainerPid(u32),
}

An entity identifier.

Variants

Global

The global entity.

Represents the root of the entity hierarchy, which is equivalent to a โglobalโ scope. This is generally used to represent a collection of metadata entries that are not associated with any specific entity, but with anything within the workload, such as host or cluster tags.

PodUid(MetaString)

A Kubernetes pod UID.

Represents the UUID of a specific Kubernetes pod.

Container(MetaString)

A container ID.

This is generally a long hexadecimal string, as generally used by container runtimes like containerd.

ContainerInode(u64)

A container inode.

Represents the inode of the cgroups controller for a specific container.

ContainerPid(u32)

A container PID.

Represents the PID of the process within a specific container.

Implementations

impl EntityId

pub fn from_local_data<S>(raw_local_data: S) -> Option<Self>
where S: AsRef<str> + Into<MetaString>,

Creates an EntityId from Local Data.

This method follows the same logic/behavior as the Datadog Agentโs origin detection logic:

If the input starts with ci-, we treat it as a container ID.

If the input starts with in-, we treat it as a container cgroup controller inode.

If the input contains a comma, we split the input and search for either a prefixed container ID or prefixed inode. If both are present, we use the container ID.

If the input starts with cid-, we treat it as a container ID.

If none of the above conditions are met, we assume the entire input is a container ID.

If the input fails to be parsed in a valid fashion (e.g., in- prefix but the remainder is not a valid integer), or is empty, None is returned.

pub fn from_pod_uid<S>(pod_uid: S) -> Option<Self>
where S: AsRef<str> + Into<MetaString>,

Creates an EntityId from a Kubernetes pod UID.

If the pod UID value is โnoneโ, this will return None.

pub fn try_into_container(self) -> Option<MetaString>

Returns the inner container ID value, if this entity ID is a Container.

Otherwise, None is returned and the original entity ID is consumed.