saluki_env::workload

Trait WorkloadProvider

pub trait WorkloadProvider {
    // Required methods
    fn get_tags_for_entity(
        &self,
        entity_id: &EntityId,
        cardinality: OriginTagCardinality,
    ) -> Option<SharedTagSet>;
    fn get_resolved_origin(
        &self,
        origin: RawOrigin<'_>,
    ) -> Option<ResolvedOrigin>;
}

Provides information about workloads running on the process host.

Required Methods

fn get_tags_for_entity( &self, entity_id: &EntityId, cardinality: OriginTagCardinality, ) -> Option<SharedTagSet>

Gets the tags for an entity.

Entities are workload resources running on the process host, such as containers or pods. The cardinality of the tags to get can be controlled via cardinality.

Returns Some(SharedTagSet) if the entity has tags, or None if the entity does not have any tags or if the entity was not found.

fn get_resolved_origin(&self, origin: RawOrigin<'_>) -> Option<ResolvedOrigin>

Resolves a raw origin.

If the origin is empty, None is returned. Otherwise, Some(ResolvedOrigin) will be returned, which contains fully resolved versions of the raw origin components.