saluki_env

Trait EnvironmentProvider

pub trait EnvironmentProvider {
    type Host: HostProvider;
    type Workload: WorkloadProvider;
    type AutodiscoveryProvider: AutodiscoveryProvider;

    // Required methods
    fn host(&self) -> &Self::Host;
    fn workload(&self) -> &Self::Workload;
    fn autodiscovery(&self) -> &Self::AutodiscoveryProvider;
}

Provides information about the environment in which the process is running.

Required Associated Types

type Host: HostProvider

Type of the host provider.

type Workload: WorkloadProvider

Type of the workload provider.

type AutodiscoveryProvider: AutodiscoveryProvider

Type of the autodiscovery provider.

Required Methods

fn host(&self) -> &Self::Host

Gets a reference to the host provider for this environment.

fn workload(&self) -> &Self::Workload

Gets a reference to workload provider for this environment.

fn autodiscovery(&self) -> &Self::AutodiscoveryProvider

Gets a reference to autodiscovery provider for this environment.